Covscan can't tell that we're not filling a buffer...
Because the pattern here is:

rc = func_call(buf, buflen)
if (rc < 0)
	return error;
buflen = rc;
buf = calloc(1, buflen)
if (!buf)
	return error;
rc = func_call(buf, buflen);

Covscan thinks the first func_call() might actually be doing something
to buf.  It isn't, but that's not obvious.  So make it NULL and 0
instead of buf and buflen on the first call.

Signed-off-by: Peter Jones <pjones@redhat.com>