Fix a potentially leaky error path.
"gcc -fanalyzer" says there's a leak of "new" from read_order().  It's a
very very very large analysis to get there, and in my experience those
are often false positives.  Nevertheless, there are some problems with
its error path - in particular, a failure to read a variable won't
currently be reported unless we allocated memory for it.

This patch changes that error path to conditionalize the error reporting
and the cleanup only on the error status, regardless of the state of
"new", and then to only free new if it was set.  This also means gcc no
longer thinks there's a leak on the error return path, even though the
"leak" only seems to happen when there was no allocation.

Signed-off-by: Peter Jones <pjones@redhat.com>